show generate failed message

+++ CodeByAI.sketchplugin/Contents/Sketch/library/generateService.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 sketchDir=$1
-apiUrl="https://service.codebyai.com/api/sketch"
+apiUrl="http://codebyai.com/api/sketch"
 cd "$sketchDir"
 if [ -f "sketch.zip" ]; then
   rm sketch.zip
@@ -11,10 +11,13 @@ fi
 zip -r sketch.zip  assets/*@2x.png artboard.json *.html.bak
 curl $apiUrl -X POST -F "sketchZip=@sketch.zip" -o CodeByAI.zip
 unzip CodeByAI.zip
-cp CodeByAI/index.html ./
-rm CodeByAI/index.html
-rm CodeByAI.zip
+if [ $? -eq 0 ]; then
+  cp CodeByAI/index.html ./
+  rm CodeByAI/index.html
+  rm CodeByAI.zip
+fi
 rm sketch.zip
+
 #if [ -f "/usr/local/bin/codebyai-cli" ]; then
 #  /usr/local/bin/codebyai-cli "$sketchDir"
 #fi
